chore: refactor main package into separate files
To make the single responsibility principle clearer, the main package is
separated into:
- main.go - the entry point
- version.go - about the version of this plugin
- upgrade_all.go - coordinates the upgrade
